Fix: 내비게이션 적용 방법 수정
