+++ src/navigations/index.js
@@ -1,6 +1,7 @@
 import { NavigationContainer } from '@react-navigation/native'
 import React from 'react'
 import MainStack from './MainStack'
+import MainTab from './MainTab'
 
 const Navigation = () => {
   return (
